Move naive attention to rocm flash attn
