+++ vllm/attention/backends/rocm_flash_attn.py
@@ -1,4 +1,5 @@
 """Attention layer ROCm GPUs."""
+import importlib
 import os
 from dataclasses import dataclass
 from typing import Dict, List, Optional, Tuple, Type
@@ -147,18 +148,24 @@ class ROCmFlashAttentionImpl(AttentionImpl):
                 f"Head size {head_size} is not supported by PagedAttention. "
                 f"Supported head sizes are: {suppored_head_sizes}.")
 
-        # NOTE: Allow for switching between Triton and CK
-        #       Defaulting to triton
+        self.use_naive_attn = _check_use_naive_attention()
+        # NOTE: Allow for switching between Triton and CK. Defaulting to triton.
         self.use_triton_flash_attn = (os.environ.get(
             "VLLM_USE_TRITON_FLASH_ATTN", "True").lower() in ("true", "1"))
-        if self.use_triton_flash_attn:
+        if self.use_naive_attn:
+            # AMD Radeon 7900 series (gfx1100) currently does not support xFormers
+            # nor FlashAttention. As a temporary workaround, we use naive PyTorch
+            # implementation of attention.
+            self.attn_fuc = _naive_attention()
+            logger.debug("Using naive attention in ROCmBackend")
+        elif self.use_triton_flash_attn:
             from vllm.attention.ops.triton_flash_attention import (  # noqa: F401
                 triton_attention)
-            self.fa_func = triton_attention
+            self.attn_func = triton_attention
             logger.debug("Using Triton FA in ROCmBackend")
         else:
             from flash_attn import flash_attn_varlen_func  # noqa: F401
-            self.fa_func = flash_attn_varlen_func
+            self.attn_func = flash_attn_varlen_func
             logger.debug("Using CK FA in ROCmBackend")
 
     def repeat_kv(self, x: torch.Tensor, n_rep: int) -> torch.Tensor:
@@ -218,26 +225,34 @@ class ROCmFlashAttentionImpl(AttentionImpl):
                 # triton attention
                 # When block_tables are not filled, it means q and k are the
                 # prompt, and they have the same length.
-                if self.use_triton_flash_attn:
+                if self.use_naive_attn or self.use_triton_flash_attn:
                     if self.num_kv_heads != self.num_heads:
                         # Interleave for MQA workaround.
                         key = self.repeat_kv(key, self.num_queries_per_kv)
                         value = self.repeat_kv(value, self.num_queries_per_kv)
-
-                    output, _ = self.fa_func(
-                        query,
-                        key,
-                        value,
-                        None,
-                        attn_metadata.seq_start_loc,
-                        attn_metadata.seq_start_loc,
-                        attn_metadata.max_prompt_len,
-                        attn_metadata.max_prompt_len,
-                        True,
-                        self.scale,
-                    )
+                    if self.use_naive_attn:
+                        output = self.attn_fuc(
+                            query,
+                            key,
+                            value,
+                            attn_metadata.prompt_lens,
+                            self.scale,
+                        )
+                    else:
+                        output, _ = self.attn_func(
+                            query,
+                            key,
+                            value,
+                            None,
+                            attn_metadata.seq_start_loc,
+                            attn_metadata.seq_start_loc,
+                            attn_metadata.max_prompt_len,
+                            attn_metadata.max_prompt_len,
+                            True,
+                            self.scale,
+                        )
                 else:
-                    output = self.fa_func(
+                    output = self.attn_func(
                         q=query,
                         k=key,
                         v=value,
@@ -282,3 +297,63 @@ class ROCmFlashAttentionImpl(AttentionImpl):
 
         # Reshape the output tensor.
         return output.view(num_tokens, hidden_size)
+
+
+def _check_use_naive_attention() -> bool:
+    # For ROCm, check whether flash attention is installed or not.
+    use_naive_attention = importlib.util.find_spec("flash_attn") is None
+    if use_naive_attention:
+        logger.warning("flash_attn is not installed. Using naive attention. "
+                       "This will take significantly more GPU memory.")
+        return True
+    return False
+
+
+def _naive_attention(
+    query: torch.Tensor,
+    key: torch.Tensor,
+    value: torch.Tensor,
+    prompt_lens: List[int],
+    scale: float,
+) -> torch.Tensor:
+    num_tokens = query.shape[0]
+    output = torch.empty_like(query)
+    start = 0
+    for _, prompt_len in enumerate(prompt_lens):
+        end = start + prompt_len
+        out = _naive_masked_attention(
+            query[None, start:end],
+            key[None, start:end],
+            value[None, start:end],
+            scale,
+        )
+        # TODO(woosuk): Unnecessary copy. Optimize.
+        output[start:end].copy_(out)
+        start += prompt_len
+
+    # Using view got RuntimeError: view size is not compatible
+    # with input tensor's size and stride (at least one
+    # dimension spans across two contiguous subspaces).
+    # Use reshape instead.
+    return output.reshape(num_tokens, -1)
+
+
+def _naive_masked_attention(
+    query: torch.Tensor,
+    key: torch.Tensor,
+    value: torch.Tensor,
+    scale: float,
+) -> torch.Tensor:
+    seq_len, _, _ = query.shape
+    attn_mask = torch.triu(torch.ones(seq_len,
+                                      seq_len,
+                                      dtype=query.dtype,
+                                      device=query.device),
+                           diagonal=1)
+    attn_mask = attn_mask * torch.finfo(query.dtype).min
+
+    attn_weights = scale * torch.einsum("qhd,khd->hqk", query, key).float()
+    attn_weights = attn_weights + attn_mask.float()
+    attn_weights = torch.softmax(attn_weights, dim=-1).to(value.dtype)
+    out = torch.einsum("hqk,khd->qhd", attn_weights, value)
+    return out
